Leanplum level test coverage

+++ LeanplumTests/Utilities/LPTestHelper.m
@@ -36,7 +36,7 @@ NSInteger DISPATCH_WAIT_TIME = 4;
     [LPAPIConfig sharedConfig].deviceId = deviceId;
 }
 
-+ (void)setupStub:(long)errorCode withFileName:(NSString *)filename {
++ (void)setupStub:(int)errorCode withFileName:(NSString *)filename {
     [OHHTTPStubs stubRequestsPassingTest:^BOOL(NSURLRequest * _Nonnull request) {
         return [request.URL.host isEqualToString:API_HOST];
     } withStubResponse:^OHHTTPStubsResponse * _Nonnull(NSURLRequest * _Nonnull request) {
